backend/pyproject.toml
[project]
name = "cronbox"
version = "0.1.0"
description = "CronBox - HTTP request scheduling service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.24.0",
    "aiosmtplib>=5.0.0",
    "alembic>=1.17.2",
    "apscheduler>=3.11.2",
    "arq>=0.26.3",
    "asyncpg>=0.31.0",
    "bcrypt>=4.2.0",
    "croniter>=6.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "orjson>=3.11.5",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "pytz>=2025.2",
    "redis>=5.3.1",
    "resend>=2.19.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.40.0",
    "yookassa>=3.9.0",
]

[project.scripts]
cronbox-worker = "app.cli:run_worker"
cronbox-scheduler = "app.cli:run_scheduler"
cronbox-server = "app.cli:run_server"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
]
